+++ KrishiBazaar/app/src/main/java/com/krishibazaar/ProfileActivity.java
@@ -378,7 +378,7 @@ public class ProfileActivity extends Fragment {
 
     private void openProductView(long productId) {
         Intent intent = new Intent(context, ProductViewActivity.class);
-        intent.putExtra(PRODUCT_ID, productId);
+        intent.putExtra(PRODUCT_ID, (int)productId);
         startActivity(intent);
     }
 }
